feat: match the scheduled Sessions row to the Messages layout
Restructure the scheduled Sessions row to mirror the Messages row: the launch
time, countdown, and Cancel now share a right-aligned group instead of an
orphaned bottom action row, and the verbose full date becomes the concise
formatClock the Messages panel already used (with the countdown gated to
pending). On narrow screens the shared time + action group wraps to a
full-width space-between line rather than jamming both against the right
margin. The two panels' duplicated formatClock / formatRelative helpers move
into a shared lib/scheduleTime module.
