+++ frontend/src/components/ScheduledMessagesPanel.tsx
@@ -5,6 +5,7 @@ import { useEffect, useState } from "react";
 import { ExpandableText } from "@/components/ExpandableText";
 import { Pager } from "@/components/Pager";
 import { usePagination } from "@/lib/usePagination";
+import { formatClock, formatRelative } from "@/lib/scheduleTime";
 import { MessageSchedule, SessionRecord } from "@/lib/types";
 
 const PAGE_SIZE = 5;
@@ -176,32 +177,3 @@ function shortSessionId(id: string): string {
   return dash >= 0 ? id.slice(dash + 1) : id;
 }
 
-function formatClock(target: Date): string {
-  const sameDay = new Date().toDateString() === target.toDateString();
-  return target.toLocaleString(undefined, {
-    month: sameDay ? undefined : "short",
-    day: sameDay ? undefined : "numeric",
-    hour: "numeric",
-    minute: "2-digit",
-  });
-}
-
-function formatRelative(target: Date): string {
-  const diff = target.getTime() - Date.now();
-  if (diff <= 0) {
-    return "any moment";
-  }
-  const minutes = Math.round(diff / 60_000);
-  if (minutes < 1) {
-    return "in <1m";
-  }
-  if (minutes < 60) {
-    return `in ${minutes}m`;
-  }
-  const hours = Math.round(minutes / 60);
-  if (hours < 48) {
-    return `in ${hours}h`;
-  }
-  const days = Math.round(hours / 24);
-  return `in ${days}d`;
-}
